% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{admm}
\alias{admm}
\title{quadratic programming with linear equality and box inequality constraints using ADMM}
\usage{
admm(Dmat, dvec, Amat, bvec, lb, ub)
}
\arguments{
\item{Dmat}{: a size (n x n) matrix representing D matrix involved in the quadratic term x. Must be positive semi-definite.}

\item{dvec}{: a size n vector representing d involved in the linear term of x.}

\item{Amat}{: a size (m x n) matrix representing A matrix involved in the linear equality constraint of x.}

\item{bvec}{: a size m vector representing b involved in the linear equality constraint.}

\item{lb}{: a size n vector representing the lower bound of box-inequality constraint of x.}

\item{ub}{: a size n vector representing the upper bound of box-inequality constraint of x.}
}
\value{
: a vector containing x, where x is a size N_co vector containing optimal weights for control units.
}
\description{
quadratic programming with linear equality and box inequality constraints using ADMM
}
