% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATT_estimate.R
\name{ATT_estimate}
\alias{ATT_estimate}
\title{Estimation of ATT trajectory}
\usage{
ATT_estimate(m, u, t, y, d, solution_w)
}
\arguments{
\item{m}{: a size (N x T) matrix of balanced panel data}

\item{u}{: (numeric) a column index for unit variable}

\item{t}{: (numeric) a column index for time variable}

\item{d}{: (numeric) a column index for treatment intervention (0 or 1)}

\item{solution_W}{: a N_ct length of vector for weights}
}
\value{
: a list containing the following variables
* ATT : (numeric) a T-dim vector of average treatment effect on treated across time (from the beginning to the end of study time)
* T_trajectory : (numeric) a T-dim vector of treatment group outcome trajectory across time
* SC_trajectory : (numeric) a T-dim vector of synthetic control outcome trajectory across time
}
\description{
Estimation of ATT trajectory
}
