% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ATTplot}
\alias{ATTplot}
\title{ATT trend graph}
\usage{
ATTplot(ATT, m, d, t, placebo)
}
\arguments{
\item{ATT}{: a (T) length vector of ATTs}

\item{m}{: a size (N x T) matrix of balanced panel data}

\item{d}{: (numeric) a column index for treatment intervention (0 or 1)}

\item{t}{: (numeric) a column index for time variable}

\item{placebo}{: a size (T X N_ct) of placebo ATTs}
}
\value{
: a graph plot
}
\description{
ATT trend graph
}
