% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{data}
\alias{data}
\title{Data preparation for optimization of V,W (.R)}
\usage{
data(m, u, t, y, d, c)
}
\arguments{
\item{m}{: a size (N x T) matrix of balanced panel data}

\item{u}{: (numeric) a column index for unit variable}

\item{t}{: (numeric) a column index for time variable}

\item{d}{: (numeric) a column index for treatment intervention (0 or 1)}

\item{c}{: a vector of column indices for time-unit-varying controls}
}
\value{
: a list containing the following variables
y0 = a (T x N_co) matrix of outcomes from control units
y1 = a (T-dim) vector of outcomes from treated units
x0_scaled = normalized x0
x1_scaled = normalized x1
z0 = a size (K x 1) matrix of control group outcomes during pre-treatment times
z1 = a (T_pre) length vector of treated group outcomes during pre-treatment times
}
\description{
Data preparation for optimization of V,W (.R)
}
