% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fn_v}
\alias{fn_v}
\title{This is the optimizing function (minimization) for V}
\usage{
fn_v(v_vec, x0_scaled, x1_scaled, z0, z1)
}
\arguments{
\item{v_vec}{: a k-dim vector of initial values for V}

\item{x0_scaled}{: a (k x Nc) matrix of scaled predictors for control group during pre-treatment times}

\item{x1_scaled}{: a (k x 1) scaled predictors for treated group during pre-treatment times}

\item{z0}{: a (T0 x Nc) matrix of outcomes for control group during pre-treatment times}

\item{z1}{: a (T0 x 1)  matrix of outcomes for treated group during pre-treatment times}
}
\value{
a scalar of V loss during pre-treatment times : t(X1-X0%*%W) %*% V %*% (X1-X0%*%W)
}
\description{
This is the optimizing function (minimization) for V
}
