% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{synthetic}
\alias{synthetic}
\title{This is the wrapping function of 1) estimation, 2) inference with parallel computing, and 3) graph}
\usage{
synthetic(y0, y1, x0_scaled, x1_scaled, z0, z1)
}
\arguments{
\item{y0}{: a (T x N_co) matrix of outcomes from control units}

\item{y1}{: a (T-dim) vector of outcomes from treated units}

\item{x0_scaled}{: a (K x N_co) matrix of predictors for control units during pre-treatment times}

\item{x1_scaled}{: a K-dim vector of predictors for treated units during pre-treatment times}

\item{z0}{: a (T0 x N_co) matrix of outcomes for control units during pre-treatment times}

\item{z1}{: a (T0-dim) vector of outcomes for treated units during pre-treatment times}
}
\value{
A list containing the following values
* ATT : (numeric) a T-dim vector of average treatment effect on treated across time (from the beginning to the end of study time)
* T_trajectory : (numeric) a T-dim vector of treatment group outcome trajectory across time
* SC_trajectory : (numeric) a T-dim vector of synthetic control outcome trajectory across time
* solution.v : (numeric) a (k x k) matrix of weight 
* solution.w : (numeric) a Nc-dim vector of synthetic weights for control units
}
\description{
This is the wrapping function of 1) estimation, 2) inference with parallel computing, and 3) graph
}
